service_a/pyproject.toml
[project]
name = "service-a"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-ninja>=1.3.0",
    "grpcio>=1.71.0",
    "grpcio-tools>=1.71.0",
]
